==> Part_2_Churn_analysis_Merge_with_outdata.ipynb <==
import random
import nltk
import pandas as pd
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.pipeline import Pipeline
from sklearn.metrics import classification_report, confusion_matrix
from sklearn.model_selection import train_test_split
from sklearn.feature_selection import SelectKBest, chi2
from sklearn.ensemble import RandomForestClassifier
import pickle
import numpy as np

pro_ser_fin=pd.read_csv('E:/Documents/Yagna/prep_data/pro_ser_fin_processed.csv')

Base_data=pd.read_csv('E:/Documents/Yagna/12th/outdata.csv')

Base_data.columns

Base_data.head()

Base_data.loc[Base_data.Ea_Number.astype(str)=='13978022826']

Transaction_Date ## renewal payment date -- 
Expiring_Contract_End_Date #2009-05-05 se these column, using thsi would be better as copared to renewal contract date

print(Base_data.X__Original_Opty_Val_Annualized.unique())
print(Base_data.X__Original_Opty_Val_Term.unique())
print(Base_data.Local_Opportunity_Currency_Code.unique())
print(Base_data['Sub-Batch_Type'].unique())
print(Base_data['First_Renewal'].unique())
print(Base_data.CON_Dollar_value.unique())
print(Base_data.Sales_Stage.unique())
                                      




Base_data.Renewal_Contract_Start_Date.unique()

Base_data[['RenewalType','Diff_inDays','Transaction_Date','Expiring_Contract_End_Date','Expiring_Month','Renewal_Contract_Duration','Renewal_Contract_Start_Date','Renewal_Contract_End_Date']]

Expiring_Contract_End_Date

Base_data_ltd=Base_data[[ 'Transaction_Date','Existing_Contract_Number',
       'Sales_Stage', 'Expiring_Contract_End_Date',
       'Expiring_Month', 'Renewal_Contract_Start_Date',
       'Renewal_Contract_End_Date', 'Renewal_Contract_Duration', 'EA_Segment',
       'Country_Name', 'State',    'X__Original_Opty_Val_Annualized',
       'Local_Opportunity_Currency_Code', 'PFST_Geo', 'PFST_Region',
        'Batch_Type', 'Sub-Batch_Type',
       'First_Renewal', 'CON_count(Existing_Contract_Number)',
	   'Diff_inDays', 'RenewalType',
       'CUS_On_Time_percent',
       'CUS_Early_percent', 'CUS_Delayed_percent', 'CON_On_Time_percent',
       'CON_Early_percent', 'CON_Delayed_percent', 'mark_1_if_renewal',
       'CON_number_renewal', 'CUS_number_renewal',
       'CON_Renewal_Contract_Duration_0.05_percentile',
       'CON_Renewal_Contract_Duration_0.1_percentile',
       'CON_Renewal_Contract_Duration_0.25_percentile',
       'CON_Renewal_Contract_Duration_0.5_percentile',
       'CON_Renewal_Contract_Duration_0.75_percentile',
       'CON_Renewal_Contract_Duration_0.9_percentile',
       'CON_Renewal_Contract_Duration_0.95_percentile',
       'CON_Renewal_Contract_Duration_0.99_percentile',
       'CON_Diff_inDays_0.05_percentile', 'CON_Diff_inDays_0.1_percentile',
       'CON_Diff_inDays_0.25_percentile', 'CON_Diff_inDays_0.5_percentile',
       'CON_Diff_inDays_0.75_percentile', 'CON_Diff_inDays_0.9_percentile',
       'CON_Diff_inDays_0.95_percentile', 'CON_Diff_inDays_0.99_percentile',
       'CUS_Renewal_Contract_Duration_0.05_percentile',
       'CUS_Renewal_Contract_Duration_0.1_percentile',
       'CUS_Renewal_Contract_Duration_0.25_percentile',
       'CUS_Renewal_Contract_Duration_0.5_percentile',
       'CUS_Renewal_Contract_Duration_0.75_percentile',
       'CUS_Renewal_Contract_Duration_0.9_percentile',
       'CUS_Renewal_Contract_Duration_0.95_percentile',
       'CUS_Renewal_Contract_Duration_0.99_percentile',
       'CUS_Diff_inDays_0.05_percentile', 'CUS_Diff_inDays_0.1_percentile',
       'CUS_Diff_inDays_0.25_percentile', 'CUS_Diff_inDays_0.5_percentile',
       'CUS_Diff_inDays_0.75_percentile', 'CUS_Diff_inDays_0.9_percentile',
       'CUS_Diff_inDays_0.95_percentile', 'CUS_Diff_inDays_0.99_percentile',
       'is_current_renewal_First_Renewal_', 'last_renewal']]

Base_data_ltd.columns=['Transaction_Date','Existing_Contract_Number', 'Sales_Stage', 'Expiring_Contract_End_Date',
       'Expiring_Month', 'Renewal_Contract_Start_Date',
       'Renewal_Contract_End_Date', 'Renewal_Contract_Duration', 'EA_Segment',
       'Country_Name', 'State', 'X__Original_Opty_Val_Annualized',
       'Local_Opportunity_Currency_Code', 'PFST_Geo', 'PFST_Region',
       'Batch_Type', 'Sub-Batch_Type', 'First_Renewal',
       'CON_Existing_Contract', 'Diff_inDays', 'RenewalType',
       'CUS_On_Time_percent', 'CUS_Early_percent', 'CUS_Delayed_percent',
       'CON_On_Time_percent', 'CON_Early_percent', 'CON_Delayed_percent',
       'mark_1_if_renewal', 'CON_number_renewal', 'CUS_number_renewal',
       'CON_Renewal_Contract_Duration_0.05_percentile',
       'CON_Renewal_Contract_Duration_0.1_percentile',
       'CON_Renewal_Contract_Duration_0.25_percentile',
       'CON_Renewal_Contract_Duration_0.5_percentile',
       'CON_Renewal_Contract_Duration_0.75_percentile',
       'CON_Renewal_Contract_Duration_0.9_percentile',
       'CON_Renewal_Contract_Duration_0.95_percentile',
       'CON_Renewal_Contract_Duration_0.99_percentile',
       'CON_Diff_inDays_0.05_percentile', 'CON_Diff_inDays_0.1_percentile',
       'CON_Diff_inDays_0.25_percentile', 'CON_Diff_inDays_0.5_percentile',
       'CON_Diff_inDays_0.75_percentile', 'CON_Diff_inDays_0.9_percentile',
       'CON_Diff_inDays_0.95_percentile', 'CON_Diff_inDays_0.99_percentile',
       'CUS_Renewal_Contract_Duration_0.05_percentile',
       'CUS_Renewal_Contract_Duration_0.1_percentile',
       'CUS_Renewal_Contract_Duration_0.25_percentile',
       'CUS_Renewal_Contract_Duration_0.5_percentile',
       'CUS_Renewal_Contract_Duration_0.75_percentile',
       'CUS_Renewal_Contract_Duration_0.9_percentile',
       'CUS_Renewal_Contract_Duration_0.95_percentile',
       'CUS_Renewal_Contract_Duration_0.99_percentile',
       'CUS_Diff_inDays_0.05_percentile', 'CUS_Diff_inDays_0.1_percentile',
       'CUS_Diff_inDays_0.25_percentile', 'CUS_Diff_inDays_0.5_percentile',
       'CUS_Diff_inDays_0.75_percentile', 'CUS_Diff_inDays_0.9_percentile',
       'CUS_Diff_inDays_0.95_percentile', 'CUS_Diff_inDays_0.99_percentile',
       'is_current_renewal_First_Renewal_', 'last_renewal']

Base_data_ltd['Renewal_Contract_Start_Date']=pd.to_datetime(Base_data_ltd.Renewal_Contract_Start_Date, format='%Y-%m-%d', errors='ignore')

pro_ser_fin['OrderDate_1']=pd.to_datetime(pro_ser_fin.OrderDate, format='%m/%d/%Y', errors='ignore')

pro_ser_fin.head()

Base_data_merge=Base_data_ltd.merge(pro_ser_fin, how='left',left_on=['Existing_Contract_Number', 'Renewal_Contract_Start_Date'], right_on=['ContractNumber', 'OrderDate'])

Base_data_merge.loc[Base_data_merge.Total_Price.isnull()==0]

71765  null, 97 match ;; 78 match 

pro_ser_fin_ContractNumber=pro_ser_fin.groupby('ContractNumber').count()[['OrderDate']]

pro_ser_fin_ContractNumber.loc[pro_ser_fin_ContractNumber.OrderDate>1].count()#6429

pro_ser_fin_ContractNumber.loc[pro_ser_fin_ContractNumber.OrderDate==1].count()#66967

pro_ser_fin_ContractNumber=pro_ser_fin_ContractNumber.loc[pro_ser_fin_ContractNumber.OrderDate==1]

pro_ser_fin_ContractNumber.reset_index(drop=False)

pro_ser_fin_fil=pro_ser_fin.merge(pro_ser_fin_ContractNumber, how='inner',on=['ContractNumber'])

pro_ser_fin_fil

pro_ser_fin.columns

Base_data_merge=Base_data_ltd.merge(pro_ser_fin_fil, how='left',left_on=['Existing_Contract_Number'], right_on=['ContractNumber'])

Base_data_ltd.shape

Base_data_merge

Base_data_merge.loc[Base_data_merge.Total_Price.isnull()==0]

Base_data_merge.loc[Base_data_merge.Total_Price.isnull()==1]

print(25087*100/71862)
print(46775*100/71862)

Base_data_merge

Base_data_merge.to_csv(r'E:/Documents/Yagna/prep_data/Base_data_merge.csv',index=False)
