==> Part_3_Churn_analysis_Smote_Scaliing_ChiSQ_RandomForest.ipynb <==
import random
import nltk
import pandas as pd
from sklearn.feature_extraction.text import TfidfVectorizer
#from sklearn.pipeline import Pipeline
from sklearn.metrics import classification_report, confusion_matrix
from sklearn.model_selection import train_test_split
from sklearn.feature_selection import SelectKBest, chi2
from sklearn.ensemble import RandomForestClassifier
import pickle
import numpy as np
from sklearn.preprocessing import MinMaxScaler
import eli5
from eli5.sklearn import PermutationImportance
import matplotlib.pyplot as plt
from imblearn.under_sampling import RandomUnderSampler
from imblearn.pipeline import Pipeline as im_Pipeline
from sklearn.pipeline import Pipeline as sk_pipeline
from imblearn.over_sampling import SMOTE

Base_data_merge=pd.read_csv('E:/Documents/Yagna/prep_data/Base_data_merge.csv')

#Base_data_merge.head()

#Base_data_merge.info()

#Base_data_merge.columns

 Base_data_merge_ltd=Base_data_merge[['Existing_Contract_Number', 'Sales_Stage',
       'Expiring_Contract_End_Date', 
       'Renewal_Contract_Duration', 'EA_Segment', 'Country_Name', 'State',
       'X__Original_Opty_Val_Annualized',
       'PFST_Geo', 'PFST_Region', 'Batch_Type', 'Sub-Batch_Type',
       'First_Renewal', 'CON_Existing_Contract', 'Diff_inDays', 'RenewalType',
       'CUS_On_Time_percent', 'CUS_Early_percent', 'CUS_Delayed_percent',
       'CON_On_Time_percent', 'CON_Early_percent', 'CON_Delayed_percent',
       'mark_1_if_renewal', 'CON_number_renewal', 'CUS_number_renewal',
       'CON_Renewal_Contract_Duration_0.05_percentile',
       'CON_Renewal_Contract_Duration_0.1_percentile',
       'CON_Renewal_Contract_Duration_0.25_percentile',
       'CON_Renewal_Contract_Duration_0.5_percentile',
       'CON_Renewal_Contract_Duration_0.75_percentile',
       'CON_Renewal_Contract_Duration_0.9_percentile',
       'CON_Renewal_Contract_Duration_0.95_percentile',
       'CON_Renewal_Contract_Duration_0.99_percentile',
       'CON_Diff_inDays_0.05_percentile', 'CON_Diff_inDays_0.1_percentile',
       'CON_Diff_inDays_0.25_percentile', 'CON_Diff_inDays_0.5_percentile',
       'CON_Diff_inDays_0.75_percentile', 'CON_Diff_inDays_0.9_percentile',
       'CON_Diff_inDays_0.95_percentile', 'CON_Diff_inDays_0.99_percentile',
       'CUS_Renewal_Contract_Duration_0.05_percentile',
       'CUS_Renewal_Contract_Duration_0.1_percentile',
       'CUS_Renewal_Contract_Duration_0.25_percentile',
       'CUS_Renewal_Contract_Duration_0.5_percentile',
       'CUS_Renewal_Contract_Duration_0.75_percentile',
       'CUS_Renewal_Contract_Duration_0.9_percentile',
       'CUS_Renewal_Contract_Duration_0.95_percentile',
       'CUS_Renewal_Contract_Duration_0.99_percentile',
       'CUS_Diff_inDays_0.05_percentile', 'CUS_Diff_inDays_0.1_percentile',
       'CUS_Diff_inDays_0.25_percentile', 'CUS_Diff_inDays_0.5_percentile',
       'CUS_Diff_inDays_0.75_percentile', 'CUS_Diff_inDays_0.9_percentile',
       'CUS_Diff_inDays_0.95_percentile', 'CUS_Diff_inDays_0.99_percentile',
       'is_current_renewal_First_Renewal_', 'last_renewal', 
       'Total_Price', 'Count_pro_ser', 'Cat_pro_ser', 'CountForCat_pro_ser',
       'Quantity_log', 'CountForCat_pro_ser_bin', 'Total_Price_log',
       'Product_price_log', 'services_price_log']]

Base_data_merge_ltd['Sales_Stage']=Base_data_merge_ltd['Sales_Stage'].map({'06 - Closed Won':0, '07 - Closed Lost':1})
Base_data_merge_ltd.Expiring_Contract_End_Date=pd.to_datetime(Base_data_merge_ltd.Expiring_Contract_End_Date, format='%Y-%m-%d')
Base_data_merge_ltd.Renewal_Contract_Duration=Base_data_merge_ltd.Renewal_Contract_Duration.astype(float)
Base_data_merge_ltd.X__Original_Opty_Val_Annualized=pd.cut(Base_data_merge_ltd.X__Original_Opty_Val_Annualized, 20).cat.codes
Base_data_merge_ltd.Renewal_Contract_Duration=pd.cut(Base_data_merge_ltd.Renewal_Contract_Duration, 20).cat.codes
Base_data_merge_ltd.First_Renewal=Base_data_merge_ltd.First_Renewal.astype(str).map({'0':False, '1':True, '2':True})

#Base_data_merge_ltd.EA_Segment.unique()

#Base_data_merge_ltd.columns

Base_data_merge_ltd.loc[Base_data_merge_ltd.last_renewal=='Did not renew',['last_renewal']]=''

Base_data_merge_ltd.drop('PFST_Region', axis=1, inplace=True)
Base_data_merge_ltd.drop('Expiring_Contract_End_Date', axis=1, inplace=True)    
Base_data_merge_ltd.drop('RenewalType', axis=1, inplace=True)   
Base_data_merge_ltd.drop('Sub-Batch_Type', axis=1, inplace=True)   
Base_data_merge_ltd.drop('mark_1_if_renewal', axis=1, inplace=True)   

dum = pd.get_dummies(Base_data_merge_ltd[['EA_Segment','Country_Name','State','PFST_Geo','Batch_Type','last_renewal','Cat_pro_ser','CountForCat_pro_ser_bin']], drop_first = True)

Base_data_merge_ltd = pd.concat([Base_data_merge_ltd, dum], axis = 1)

# Now let's see the head of our dataframe.
#housing.head()
# Drop 'furnishingstatus' as we have created the dummies for it
Base_data_merge_ltd_1hot=Base_data_merge_ltd.drop(['EA_Segment','Country_Name','State','PFST_Geo','Batch_Type','last_renewal','Cat_pro_ser','CountForCat_pro_ser_bin'], axis = 1)

#housing.head()

#Base_data_merge_ltd.info()

#Base_data_merge_ltd_null=Base_data_merge_ltd_1hot.isnull().sum()

#Base_data_merge_ltd_1hot.loc[Base_data_merge_ltd_1hot>0]

#Base_data_merge_ltd_clean.isnull().sum()

Base_data_merge_ltd_clean=Base_data_merge_ltd_1hot.dropna()

Base_data_merge_ltd_clean.shape

# check version number
#import imblearn
#print(imblearn.__version__)

#from imblearn.over_sampling import SMOTE
#oversample = SMOTE()
#X, y = oversample.fit_resample(X, y)

Y = Base_data_merge_ltd_clean.Sales_Stage
X =Base_data_merge_ltd_clean.drop('Sales_Stage', axis=1)

X_train, X_test, y_train, y_test = train_test_split(X, Y, test_size=0.2,random_state=9)

X_train_Existing_Contract_Number=X_train.Existing_Contract_Number                         
X_test_Existing_Contract_Number=X_test.Existing_Contract_Number   

X_train =X_train.drop('Existing_Contract_Number', axis=1)
X_test =X_test.drop('Existing_Contract_Number', axis=1)


oversample = SMOTE()
X_train, y_train = oversample.fit_resample(X_train, y_train)

from itertools import product

def expand_grid(dictionary):
   return pd.DataFrame([row for row in product(*dictionary.values())], 
                       columns=dictionary.keys())

dictionary = {'over': [.2,.3,.4,.5,.6,.7,.8,.9], 
              'under': [.5,.6,.7,.8,.9]
              }
dictionary=expand_grid(dictionary)
#dictionary
#dictionary.ngram_range[0]
#dictionary.shape[0]
#dictionary

#dictionary

result_set

result_set = []
sk_pipeline = sk_pipeline([('sca',MinMaxScaler()),
                     ('chi',  SelectKBest(chi2, k=11)),
                     ('clf', RandomForestClassifier())])
ytest = np.array(y_test)
for i in range (0,dictionary.shape[0],1):
    print(i)
    print(dictionary.over[i])
    print(dictionary.under[i])
    over = SMOTE(sampling_strategy=dictionary.over[i])
    under = RandomUnderSampler(sampling_strategy=dictionary.under[i])
    steps = [('o', over), ('u', under)]
    #steps = [('o', over)]
    pipeline = Pipeline(steps=steps)
    # transform the dataset
    X_train_gen, y_train_gen = pipeline.fit_resample(X_train, y_train)
    model = sk_pipeline.fit(X_train_gen, y_train_gen)
    result_set.extend([i,accuracy_score(ytest, model.predict(X_test))])    

over = SMOTE(sampling_strategy=.5)
under = RandomUnderSampler(sampling_strategy=.5)
steps = [('o', over), ('u', under)]
    #steps = [('o', over)]
im_pipeline = im_Pipeline(steps=steps)
    # transform the dataset
X_train_gen, y_train_gen = im_pipeline.fit_resample(X_train, y_train)

#X_train.
#X_testgroupby('ContractNumber').count()
print(y_train.count())
print(y_test.count())
print(y_train.sum())
print(y_test.sum())
print(y_train.sum()/y_train.count())
print(y_train_gen.sum()/y_train_gen.count())
print(y_test.sum()/y_test.count())

X_train_gen=X_train_gen[['CON_Delayed_percent','CUS_Delayed_percent','First_Renewal','Country_Name_INDIA','CON_number_renewal','Country_Name_MALAYSIA',
'State_Maharashtra','Country_Name_SINGAPORE','State_New South Wales','PFST_Geo_RENEWALS_TEMPORARY','PFST_Geo_APAC','State_Victoria',
'State_Tamil Nadu','Country_Name_AUSTRALIA']]
X_test=X_test[['CON_Delayed_percent','CUS_Delayed_percent','First_Renewal','Country_Name_INDIA','CON_number_renewal','Country_Name_MALAYSIA',
'State_Maharashtra','Country_Name_SINGAPORE','State_New South Wales','PFST_Geo_RENEWALS_TEMPORARY','PFST_Geo_APAC','State_Victoria',
'State_Tamil Nadu','Country_Name_AUSTRALIA']]

Sk_pipeline = sk_pipeline([('sca',MinMaxScaler()),
                     ('chi',  SelectKBest(chi2, k=14)),
                     ('clf', RandomForestClassifier())])

model = Sk_pipeline.fit(X_train_gen, y_train_gen)
with open('Churn_chi_bal_Only14_33O_ltd.pickle', 'wb') as f:
    pickle.dump(model, f)
ytest = np.array(y_test)
# confusion matrix and classification report(precision, recall, F1-score)
print(classification_report(ytest, model.predict(X_test)))
print(confusion_matrix(ytest, model.predict(X_test)))


from sklearn.metrics import accuracy_score

threshold = 0.9

predicted_proba = model.predict_proba(X_test)
predicted = (predicted_proba [:,1] >= threshold).astype('int')

accuracy = accuracy_score(y_test, predicted)
accuracy

print(classification_report(ytest, predicted))
print(confusion_matrix(ytest, predicted))

#specificity=356/(356+556)
#recall for 0 
356/(356+52) #quality
356/(356+556)# quantity

#10
  precision    recall  f1-score   support

           0       0.92      0.98      0.95      6678
           1       0.76      0.41      0.53       912

    accuracy                           0.91      7590
   macro avg       0.84      0.69      0.74      7590
weighted avg       0.90      0.91      0.90      7590

#5
                 precision    recall  f1-score   support

           0       0.92      0.99      0.96      6678
           1       0.87      0.39      0.54       912

    accuracy                           0.92      7590
   macro avg       0.90      0.69      0.75      7590
weighted avg       0.92      0.92      0.91      7590

[[6626   52]
 [ 556  356]]
#6
  precision    recall  f1-score   support

           0       0.92      0.98      0.95      6678
           1       0.78      0.40      0.53       912

    accuracy                           0.91      7590
   macro avg       0.85      0.69      0.74      7590
weighted avg       0.91      0.91      0.90      7590

[[6576  102]
 [ 550  362]]

y_pred=model.predict(X_test)

y_pred

y_test

mask = model['chi'].get_support()
new_features = X_train.columns[mask]

cols = model['chi'].get_support(indices=True)
print(cols)
features_df_new = X_train.iloc[:,cols]
features_df_new.columns

model['chi'].info()

scores = pd.DataFrame(-np.log10(model['chi'].pvalues_))

predictors=pd.DataFrame(X_train.columns)

##change didn't renew to null -- Did not renew ex contacrt 3777541929
## doesn't make sence to have not renewal  as a valid entry for past renewal

#Sub-Batch_Type_WINBACK=86.72553029, last_renewal_Did not renew=108.92708617

col_res=pd.concat([predictors,scores], axis = 1)

col_res.columns=['Variable','Score']

col_res.sort_values('Score', inplace=True, ascending=False)

['CON_Delayed_percent','CUS_Delayed_percent','First_Renewal','Country_Name_INDIA','CON_number_renewal','Country_Name_MALAYSIA',
'State_Maharashtra','Country_Name_SINGAPORE','State_New South Wales','PFST_Geo_RENEWALS_TEMPORARY','PFST_Geo_APAC','State_Victoria',
'State_Tamil Nadu','Country_Name_AUSTRALIA']

col_res.head(14)

plt.figure(figsize=(20,10))
plt.bar(range(len(predictors)), scores)
plt.xticks(range(len(predictors)), predictors, rotation='vertical')
plt.show()

model['chi'].shape()

model['chi'].pvalues_

model['chi'].scores_

model['clf'].feature_importances_

list(zip(train[features], model['clf'].feature_importances_))

rf = RandomForestClassifier()
perm = PermutationImportance(model['clf'], cv = None, refit = False, n_iter = 50).fit(X_train, y_train)
perm_imp_eli5 = imp_df(X_train.columns, perm.feature_importances_)
